FEAT: 뉴스 접근 기록 생성 요청을 위해 필요한 RequestCreateNewsAccessLogDTO 생성
